% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{class}
\name{Endpoint-class}
\alias{Endpoint-class}
\alias{Endpoint}
\title{Endpoint}
\description{
Executes http requests against the specified URL.
}
\section{Fields}{

\describe{
\item{\code{base_url}}{character. Base URL.}
}}


