% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getdocument}
\alias{getdocument}
\title{Retrieves document content as HTML string}
\usage{
getdocument(document_id)
}
\arguments{
\item{document_id}{Document id}
}
\value{
character
}
\description{
Retrieves document content as HTML string
}
\examples{
document <- get_document("GZ01MJU21")
}
\references{
https://www.riksdagen.se/sv/dokument-och-lagar/riksdagens-oppna-data/anvandarstod/sa-fungerar-dokument-id/
}
