% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heleramcar5-package.R
\docType{package}
\name{heleramcar5-package}
\alias{heleramcar5}
\alias{heleramcar5-package}
\title{heleramcar5: Riksdagen SDK}
\description{
SDK for Riksdagen database.
}
\author{
\strong{Maintainer}: Helena Llorens \email{hllor282@student.liu.se}

Authors:
\itemize{
  \item Carmen Rodríguez \email{carro162@student.liu.se}
  \item Ramón Amézquita \email{ramam120@student.liu.se}
}

}
\keyword{internal}
