% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{queryDocs}
\alias{queryDocs}
\title{Searches documents in the Riksdagen database}
\usage{
queryDocs(query)
}
\arguments{
\item{query}{Search query (e.g., "budget").}
}
\value{
data.frame
}
\description{
Searches documents in the Riksdagen database
}
\examples{
result <- queryDocs(query = "budget")
}
