% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{searchdocs}
\alias{searchdocs}
\title{Searches documents in the Riksdagen database.}
\usage{
searchdocs(query)
}
\arguments{
\item{query}{Search query (e.g., "budget").}
}
\value{
data.frame
}
\description{
Searches documents in the Riksdagen database.
}
\examples{
result <- searchdocs(query = "budget")
}
