% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{wordsFrequency}
\alias{wordsFrequency}
\title{Auxiliary function for counting the number of words in a string.}
\usage{
wordsFrequency(text)
}
\arguments{
\item{text}{Input text.}
}
\value{
data.frame
}
\description{
Auxiliary function for counting the number of words in a string.
}
