% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{getDocument}
\alias{getDocument}
\title{Retrieves document content as HTML string}
\usage{
getDocument(document_id)
}
\arguments{
\item{document_id}{Document id}
}
\value{
character
}
\description{
Retrieves document content as HTML string
}
\examples{
document <- getDocument(document_id = "sfs-2016-1091")

}
\references{
https://www.riksdagen.se/sv/dokument-och-lagar/riksdagens-oppna-data/anvandarstod/sa-fungerar-dokument-id/
}
